npc and heretic in progress
